You said : Other classes (Barbarian, Bard, Cleric, Druid, Paladin, Rogue, Sorcerer) do not typically grant specific, pre-determined feats from the general feat list as part of their core progression. They have class features and special abilities, some of which are very powerful (like Rage, Bardic Music, Turn Undead, Wild Shape, Smite Evil, Sneak Attack, Spellcasting), but these are not "feats" in the sense of items taken from the feat list.
In fact, they are feats. Like you said, class "features". Features are feats. What you call special ability is a type of feat. remember when i asked you to add a type for the feat structure ? you must use it now for this task.

for example, at level 1, a druid received the feat Animal Companion (Ex), which is a Extraordinaty feat. This kind of class granted feat should not appear in the add feat list of course.

now that you have the information, improve the dnd-classes.json and dnd-feats.json to include all grantedFeats from the d&d 3.5 srd for all levels for all classes. it is important that the dnd-classes.json and dnd-feats.json are extensive and complete. use the type field in the feat.