here is what is display in feat list :
Mind Blade (Su)(Supernatural Ability)(1d6, +0 enh) (Soulknife)
As a move action, a soulknife can create a semisolid blade composed of psychic energy distilled from his own mind. The blade is identical in all ways (except visually) to a short sword of a size appropriate for its wielder. The blade gains enhancement bonuses as the soulknife levels.
Prerequisites: None

1. (Su)(Supernatural Ability) is redundant. could you hide (Su) or remove it ? keep (Supernatural Ability) but remove parenthesis as this text is already in a pill.
2. don't use abbreviation in the note : +0 enh for example. Expand the full word and capitalize it.
3. (Soulknife) should be a badge before the title of the feat (Mind Blade), and without parenthesis as it will be in a badge.

This formatting should, of course, be the case for all feats in the list.